Tavern Keeper
Initializing from prefab is not working, so I will have the party members start out within the scene and only follow the player when they are bought. Shopkeeper can be made easily but has to be deeply integrated with the inventory system.  Party Members need to search for enemies and attack them. Or they need to act as a shield that the enemies fight against first before they go after the player.